Update README.md to include Brink Flair 325
Hi Lukas,

Thanks so much for developing this. Connecting to my Brink Flair 325 was a breeze (pun intended) and I couldn't find a way to email you so instead I thought I'd make a PR to show that it works for the Brink 325 (At least I can change the mode from HA, which is the most important thing for me!).
I may be back at some point once I've played around with your code to see if I can help add any functionality but that is not a promise!